now i am returning html data as ajax.php
